Refactoring, add CLI
